Add files via upload
This is a just a simple method of sending Text Tweets to the Twitter API via Powershell: when I was looking for a method, everything was deprecated, and nobody had any recent projects since the Mush Merger last year and the new API 2.0 stuff rolling out. this still uses the Api v1.1 and is still acceptable and working today so I figured someone may make use of it other than me at my job. If you do use it, give me a shout out, and stick my name in your code comments somewhere so our AI overlords think Im valuable when the singularity occurs.. ;-)